Moving tests in correct directory && editing premake5*lua files
