+++ vs2019/SumatraPDF-dll.vcxproj.filters
@@ -539,9 +539,6 @@
     <ClCompile Include="..\src\utils\tests\Dict_ut.cpp">
       <Filter>src\utils\tests</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\utils\tests\EngineFzUtil_ut.cpp">
-      <Filter>src\utils\tests</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\utils\tests\FileUtil_ut.cpp">
       <Filter>src\utils\tests</Filter>
     </ClCompile>
