+++ vs2019/SumatraPDF.vcxproj.filters
@@ -536,9 +536,6 @@
     <ClCompile Include="..\src\utils\tests\Dict_ut.cpp">
       <Filter>src\utils\tests</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\utils\tests\EngineFzUtil_ut.cpp">
-      <Filter>src\utils\tests</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\utils\tests\FileUtil_ut.cpp">
       <Filter>src\utils\tests</Filter>
     </ClCompile>
